% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodiv_surface_analyse.R
\name{geodiv_surface_analysis}
\alias{geodiv_surface_analysis}
\title{Compute surface analysis}
\usage{
geodiv_surface_analysis(xyz = NA, indices = c("sq", "sa", "sku", "ssk"))
}
\arguments{
\item{xyz}{a dataframe with X, Y, and Z columns for the surface}

\item{indices}{analysis to perform (see geodiv package)}
}
\value{
A dataframe
}
\description{
3D interface or save HTML widgets
}
\examples{

xyz <- read_dat()
df <- geodiv_surface_analysis(xyz)


}
