% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_surface_3d.R
\name{plotly_surface_3d}
\alias{plotly_surface_3d}
\title{Create a 3D surface with Plotly}
\usage{
plotly_surface_3d(
  xyz = NA,
  scale = "µm",
  markers = NA,
  size.markers = 10,
  color.markers = "#FF0000",
  opacity.markers = 0.3,
  export.plot = TRUE,
  xyz.name = "surf3d.html",
  dirOut = "C:/Rprojects/Rdev/3dlithic/results/"
)
}
\arguments{
\item{xyz}{a dataframe with X, Y, and Z columns for the surface}

\item{scale}{the scale of the axis (mm, cm, etc.)}

\item{size.markers, color.markers, opacity.markers}{grahical parameters}

\item{export.plot}{if TRUE export, if FALSE display}

\item{xyz.name}{name of the output plot. Only useful when export.plot is TRUE}

\item{dirOut}{name of the output folder. Only useful when export.plot is TRUE}

\item{marker}{a dataframe with X, Y, and Z columns of additional points. NA by default}
}
\value{
A dataframe
}
\description{
3D interface or save HTML widgets
}
\examples{

xyz <- read_dat()
plotly_surface_3d(xyz, export.plot = FALSE)

}
