% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dat.R
\name{read_dat}
\alias{read_dat}
\title{Create a map, whether static or interactive leaflet, from a GeoJSON file}
\usage{
read_dat(
 
    file.path = "https://raw.githubusercontent.com/zoometh/Rdev/master/3dlithic/inst/extdata/confoc_3d.dat",
  col.xyz = c(1, 2, 3),
  xyz.write = FALSE,
  xyz.name = "confoc_3d.xyz",
  dirOut = "C:/Rprojects/Rdev/3dlithic/inst/extdata/"
)
}
\arguments{
\item{col.xyz}{indices of the X, Y, and Z columns to read}

\item{xyz.write}{is TRUE, write a TSV of these columns}

\item{xyz.name}{name of the TSV output. Only useful if xyz.write is TRUE}

\item{dirOut}{name of the output folder. Only useful if xyz.write is TRUE}

\item{map.name}{the path to a .dat from the ConFocal}
}
\value{
A dataframe
}
\description{
clean a .dat file coming from the Confocal and export it to .xyz
}
\examples{

xyz <- read_dat()

}
